+++ src/utils/utils.js
@@ -9,6 +9,13 @@ function debug(message, data) {
 }
 
 function loadSettings() {
+  const DEFAULT_SETTINGS = {
+    domains: ['hmartos.github.io'],
+    selector: 'button.btn-show-tooltip',
+    tooltipText: 'WARNING',
+    defaultSettings: true,
+  };
+
   return new Promise((resolve, reject) => {
     chrome.storage.sync.get(['warningTooltipOptions'], function(result) {
       try {
@@ -17,6 +24,11 @@ function loadSettings() {
           throw new Error('settings-not-loaded');
         }
         const settings = result.warningTooltipOptions;
+        if (!settings) {
+          debug('Using default settings');
+          resolve(DEFAULT_SETTINGS);
+          return;
+        }
         debug('Loaded settings', settings);
         resolve(settings);
       } catch (error) {
@@ -27,11 +39,30 @@ function loadSettings() {
   });
 }
 
-function isAllowedDomain(domains, hostname) {
-  // TODO Allow regular expresions in domains
+function saveSettings(settings) {
+  return new Promise((resolve, reject) => {
+    chrome.storage.sync.set({ warningTooltipOptions: settings }, function() {
+      try {
+        if (chrome.runtime.lastError) {
+          console.error('Error loading settings', chrome.runtime.lastError.message);
+          throw new Error('settings-not-saved');
+        }
+        debug('Saved settings', settings);
+        resolve();
+      } catch (error) {
+        console.error('Error loading settings', error);
+        reject(error);
+      }
+    });
+  });
+}
+
+function isAllowedDomain(url, domains) {
+  const hostname = new URL(url).hostname;
+
   if (
     _.find(domains, domain => {
-      return domain === hostname || `www.${domain}` === hostname;
+      return new RegExp(domain, 'i').test(hostname);
     })
   ) {
     debug(`Visited page with hostname '${hostname}' in domain list`, domains);